pyproject.toml
[tool.poetry]
name = "blog"
version = "0.1.0"
description = "Personal Blog"
authors = ["Pbezerra-dev <herbetyp@gmail.com>"]
license = "AGPL-3.0-only"

[tool.poetry.dependencies]
python = "^3.7"
django_summernote = "^0.8.11"
pillow = "^7.1.2"
django-crispy-forms = "^1.9.0"
requests = "^2.23.0"
django = "2.2.7"
python-decouple = "^3.3"
dj-database-url = "^0.5.0"
django-axes = "^5.3.1"

[tool.poetry.dev-dependencies]
flake8 = "^3.7.9"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
